pyproject.toml
[project]
name = "x32-recorder"
version = "0.1.0"
description = ""
authors = [
    { name = "Tobias Reineke", email = "tobi@g3th.net" }
]
dependencies = [
    "django>=4.1.2",
    "djangorestframework>=3.15.2",
    "gunicorn>=20.1.0",
    "numpy>=1.24.4",
    "sounddevice>=0.5.2",
    "whitenoise>=6.7.0",
]
requires-python = ">=3.8"

[tool.uv]
dev-dependencies = [
    "black>=22.10.0",
]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
